requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
et_xmlfile==2.0.0
fuzzywuzzy==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.0
numpy==2.2.6
ollama==0.4.8
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
